added products.json file and refactor routes

+++ src/main/java/com/support/aninfosupportmodule/repository/TicketRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(path = "tickets", collectionResourceRel = "tickets")
-public interface TicketRepository extends CrudRepository<Ticket, Integer> {
+public interface TicketRepository extends CrudRepository<Ticket, Long> {
 }
